backend/requirements.txt
# RiskAdvisor Backend
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Data & ML
pandas>=2.2.0
numpy>=2.0.0
scikit-learn>=1.5.0
xgboost>=2.0.0

# Optimization
pulp==2.8.0
cvxpy>=1.5.0
scipy>=1.12.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# HTTP & AeroRisk Integration
requests>=2.31.0
httpx>=0.26.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
loguru>=0.7.0

# NLP (for Executive Interface)
transformers>=4.37.0
torch>=2.1.0
